Refactor: 기존 Post Controller, Service, UseCase를 Register 전용으로 변경
