+++ src/main/java/com/service/dida/domain/post/dto/EditPostRequestDto.java
@@ -6,13 +6,12 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
-
 @Getter
 @Builder
 @AllArgsConstructor
-public class PostPostReq {
-    @NotBlank(message = "게시글을 작성하려는 nftId를 입력하세요.")
-    private Long nftId;
+public class EditPostRequestDto {
+    @NotBlank(message = "수정하려는 게시글의 Id를 입력하세요.")
+    private Long postId;
     @NotBlank(message = "제목을 입력하세요.")
     @Size(min = 2, max = 30, message = "제목의 길이는 2~20이어야 합니다.")
     private String title;
