+++ src/main/java/com/service/dida/domain/post/usecase/RegisterPostUseCase.java
@@ -1,0 +1,7 @@
+package com.service.dida.domain.post.usecase;
+
+import com.service.dida.domain.post.dto.PostPostRequestDto;
+
+public interface RegisterPostUseCase {
+    void createPost(Long memberId, PostPostRequestDto postPostRequestDto);
+}
